pyproject.toml
[project]
name = "events-app-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jimmy Martinez",email = "jimmymartinez591@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "events_app_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
